pyproject.toml
[project]
name = "autoscout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.16",
    "fastapi[standard]>=0.135.3",
    "fastembed>=0.8.0",
    "google-genai>=1.73.0",
    "mcp>=1.27.0",
    "openai>=2.31.0",
    "openinference-instrumentation-agno>=0.1.30",
    "opentelemetry-api>=1.41.0",
    "opentelemetry-sdk>=1.41.0",
    "pgvector>=0.4.2",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.0",
    "sqlalchemy>=2.0.49",
    "yfinance>=1.2.2",
]
